GDB: fixed thread sometimes switch to other when trying pause thread

+++ src/dbg/debugger.cpp
@@ -403,7 +403,10 @@ static bool getConditionValue(const char* expression)
 
 void cbPauseBreakpoint()
 {
-    hActiveThread = ThreadGetHandle(((DEBUG_EVENT*)GetDebugData())->dwThreadId);
+    if(hActiveThread != ThreadGetHandle(((DEBUG_EVENT*)GetDebugData())->dwThreadId))
+    {
+        return;
+    }
     auto CIP = GetContextDataEx(hActiveThread, UE_CIP);
     DeleteBPX(CIP);
     GuiSetDebugState(paused);
